requirements.txt
# RSS / scraping
feedparser>=6.0
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0

# AI script generation
anthropic>=0.40

# Video editing
moviepy>=2.0
Pillow>=10.0

# Scheduling
apscheduler>=3.10

# Configuration
pyyaml>=6.0
python-dotenv>=1.0

# Instagram publishing
requests>=2.31

# Dashboard
streamlit>=1.35

# Utilities
tenacity>=8.2
pydantic>=2.5
